+++ zenovis/include/zenovis/Scene.h
@@ -10,14 +10,9 @@ namespace zenovis {
 struct Camera;
 struct DrawOptions;
 struct ShaderManager;
-struct EnvmapManager;
 struct GraphicsManager;
-struct LightCluster;
-struct Light;
 struct IGraphic;
 struct IGraphicDraw;
-struct DepthPass;
-struct ReflectivePass;
 
 namespace opengl {
 struct VAO;
@@ -26,26 +21,16 @@ struct VAO;
 struct Scene : zeno::disable_copy {
     std::unique_ptr<Camera> camera;
     std::unique_ptr<DrawOptions> drawOptions;
-    std::unique_ptr<LightCluster> lightCluster;
     std::unique_ptr<ShaderManager> shaderMan;
-    std::unique_ptr<EnvmapManager> envmapMan;
     std::unique_ptr<GraphicsManager> graphicsMan;
     std::vector<std::unique_ptr<IGraphicDraw>> hudGraphics;
-    std::unique_ptr<DepthPass> mDepthPass;
-    std::unique_ptr<ReflectivePass> mReflectivePass;
     std::unique_ptr<opengl::VAO> vao;
 
     Scene();
     ~Scene();
 
-    void fast_paint_graphics();
-    void drawSceneDepthSafe(bool reflect, bool isDepthPass);
-    void my_paint_graphics(int samples, bool isDepthPass);
-    void draw_small_axis();
+    void draw();
     std::vector<char> record_frame_offline(int hdrSize = 1, int rgbComps = 3);
-    bool anyGraphicHasMaterial();
-
-    void draw(unsigned int target_fbo = 0);
 };
 
 } // namespace zenovis
