+++ zenovis/include/zenovis/Session.h
@@ -25,8 +25,6 @@ struct Session : zeno::disable_copy {
     void set_window_size(int nx, int ny);
     void set_curr_frameid(int frameid);
     int get_curr_frameid();
-    /* float get_solver_interval(); */
-    /* float get_render_fps(); */
     void set_show_grid(bool flag);
     void look_perspective(float cx, float cy, float cz, float theta,
                           float phi, float radius, float fov,
@@ -41,17 +39,6 @@ struct Session : zeno::disable_copy {
     void set_smooth_shading(bool smooth);
     void set_normal_check(bool check);
     void set_render_wireframe(bool render_wireframe);
-    /* unsigned int setup_env_map(std::string name); */
-    /* void setLightData(int index, std::tuple<float, float, float> dir, */
-    /*                   float height, float softness, */
-    /*                   std::tuple<float, float, float> tint, */
-    /*                   std::tuple<float, float, float> color, float intensity); */
-    /* int getLightCount(); */
-    /* void addLight(); */
-    /* std::tuple<std::tuple<float, float, float>, float, float, */
-    /*            std::tuple<float, float, float>, std::tuple<float, float, float>, */
-    /*            float> */
-    /* getLight(int i); */
 };
 
 } // namespace zenovis
